services/market-pulse-scanner/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
asyncpg==0.29.0
supabase==2.0.2
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
kafka-python==2.0.2
celery==5.3.4
redis==5.0.1
beautifulsoup4==4.12.2
python-multipart==0.0.6
python-dotenv==1.0.0
sentry-sdk[fastapi]==1.38.0
schedule==1.2.0
aiofiles==23.2.1
pandas==2.0.3
numpy==1.24.3
textblob==0.17.1
vaderSentiment==3.3.2
tweepy==4.14.0
linkedin-api==2.0.0a5